Finished backend code pulling to main, all features fixed and added (#34)
* Users php is done on index

implemented CAS and users are now added to database. note: this needs to be implemented on all pages

* Moved pages

Moved location of pages to be in root directory and changed file paths accordingly.

* Small changes

Did a few small changes: added icon and had logo redirect to the index

* Added Dynamic Categories

Added php and js to dynamically populate categories and asynchronously populate subcategories based on user selected category/subcategory.

* I h8 php syntax

I did the query and form stuff to add an item there is a PDO exception which states there is an SQL syntax error in the execute. I will continue to work on this but posted in case someone can figure it out first.

* Uploading to the database is now possible

Found the SQL error description is an SQL keyword so had to change its name in the DB which solved this issue. Just had to walk away for a few hours and look at it from a different angle.

* Added Delete item

* Update to formatting seller dash

* changed query

* Finished Delete confirmation

Delete now has a confirmation.

* images work lol

* edit functionality added (#30)

* Update to index

items are now dynamically loaded with php

* New functionality

Able to view items on the home page and view items on product page.

* MAde more fducntioas

* Thuloa backend 2 pull all updates (#31)

* admin functionality added

* banning users implemented, as well as admin controls

* edit form functionality updated

* browse page initialized, displays all items

* Index done

* index functionality added

* filtering

* Search and filter

yaasss queen

* Thuloa backend 3 (#32)

* changed filepaths

* fixed buyer dashboard

* allowed for bigger rcsids

* fixed image justification

* Update search

yasssss

* search maybe

* dsdsdsd

* Update browse.php

* Update browse.php

* fully functional browse page

used get with forms to make full functional browse page

* Update stuff

* Fixed spacing issues, directories, and offer glitch

* minor similar items change

* another minor change to similar items display

* edit listing fixed

* directories changed

Co-authored-by: CianHowell <howelc@rpi.edu>